pyproject.toml
[project]
name = "drbnatell"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi[standard]>=0.135.1",
    "ffmpy>=1.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pillow>=12.1.1",
    "pilmoji>=2.0.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "tortoise-orm>=1.1.6",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
